Remove fallback for `op(α::Number, L::Operator)`
This PR removes (unintended?) fallbacks for `op(α, L)` and `op(L, α)` where `op ∈ (:*, :/, :\)` when `α` is a `Number` and `L` an `AbstractDiffEqLinearOperator`. 

This is because, currently, DiffEqOperators overwrites (pirates) at least the `α * L` method to create a lazy `DiffEqScaledOperator`. My understanding from Slack discussions with Chris is that these operations with scalars should be removed here to avoid downstream piracy from DiffEqOperators which provides more efficient lazy operator constructions.