Throw early warning function argument errors
Note that there is a broken case where `f` is `iip` and `jac` has `oop` and `iip` dispatches, but it throws a non-conforming error. That will take quite a bit of code to solve and is quite an edge case, so I'm not going to let that block this and will just keep as `@test_broken` with an issue open.