Allow for early wrapping
This is mostly for testing.

```julia
using OrdinaryDiffEq, SnoopCompile, Profile, ProfileView
function lorenz(du, u, p, t)
    du[1] = 10.0(u[2] - u[1])
    du[2] = u[1] * (28.0 - u[3]) - u[2]
    du[3] = u[1] * u[2] - (8 / 3) * u[3]
end

@time begin
    lorenzprob = ODEProblem{true, SciMLBase.AutoSpecialize}(lorenz, [1.0; 0.0; 0.0], (0.0, 1.0), Float64[])
    sol = solve(lorenzprob, Rosenbrock23())
end

# FunctionWrapperSpecialize:
# 1.475326 seconds (83.83 k allocations: 3.442 MiB, 99.79% compilation time)
# 0.000184 seconds (458 allocations: 40.070 KiB)
# AutoSpecialize:
# 1.597643 seconds (958.02 k allocations: 49.979 MiB, 99.85% compilation time)
# 0.000182 seconds (467 allocations: 40.203 KiB
```

This is with a change to OrdinaryDiffEq.jl precompiling both. Currently it looks like this does not give any advantage (other than in allocations) over using the AutoSpecialize form. But it's good to keep this machinery around because it's completed and the speed of light version of such wrapping, so it can allow us to easily build fully complete pre-wrapped system image packages while not impacting the standard use case.

This also adds a specialization fallback to FullSpecialize on the special AbstractSciMLFunctions, which they will not want to auto-wrap.