Safely update retcodes to enums
I think I found a way that is relatively safe. In all of our code throughout SciML, I noticed that the only thing that was truly done was `sol.retcode == :Success`, etc. in which case a relatively small amount of functions put over a retcode enum could be non-breaking to all functional uses of the previous retcode. If this is the case, we may finally have a good way to upgrade the one piece I find most hideous in the interface.