Rename the extension packages and do RCall as well
Fixes the name of the extensions so they don't clash elsewhere. Also adds an RCall one which is basic and assumes out of place, which is a fine assumption given that R is copy-on-write and thus mutation of the form of an in-place function is just not possible. This means that we don't get the great error messages, but oh well we can figure that out later.

Fixes https://github.com/SciML/diffeqr/issues/41

Better version requires https://github.com/JuliaInterop/RCall.jl/issues/503