Add Catalyst as a downstream test
I'm not sure I've done this right, but I'd like to propose adding Catalyst as a downstream test. We have very comprehensive tests, including for all kinds of symbolic indexing and plotting courtesy of @TorkelE, and I think some of the issues with stuff inadvertently getting broken might get caught if Catalyst was tested on such PRs.