fix: reorder `SplitFunction` fields to avoid breaking change
The existing syntax is `initializeprob, ..., nlprob`. Trying to add
`initialization_data` in the middle breaks the non-kwarg-only method.
Putting it at the end fixes this issue. Anything old still has the order
it relies on, with `initialization_data` defaulting to `nothing`, and
anything new would just have to provide the redundant kwargs if it needs
to specify `initialization_data`.